Days end commit. Lots of refactoring and cleaning up. Removed unthreaded write altogether. Still having some write bugs. TODO: Solve writebug and implement a buffer in huge tables. Added init for huge tables

+++ aiowriter.h
@@ -7,8 +7,8 @@
 //http://stackoverflow.com/questions/8629690/linux-async-io-with-libaio-performance-issue
 
 int aiow_init(struct opt_s *opt, struct recording_entity * re);
-int aiow_write(struct recording_entity * re, void * start, size_t count);
-int aiow_check(struct recording_entity * re);
+long aiow_write(struct recording_entity * re, void * start, size_t count);
+long aiow_check(struct recording_entity * re);
 int aiow_close(struct recording_entity * re, void * stats);
 int aiow_wait_for_write(struct recording_entity * re);
 int aiow_init_rec_entity(struct opt_s * opt, struct recording_entity * re);
